Add files via upload
How do we understand visual information? Gestalt is a collection of principles to help us decipher the meaning of images. I created this simple infographic to illustrate Gestalt principles. I'm interested in creating visual legal information so the law is easier to understand, perhaps even if you can't read. It's an incredible challenge.